refactor: Removed artifacts for a merge
