+++ .gitignore
@@ -12,4 +12,4 @@ compile_commands.json
 run_tests
 doxy/
 lib/
-log_valgrind/
+logs/
