+++ Makefile
@@ -4,7 +4,7 @@ ARGS		?=	config/example.conf
 HDIR		=	inc
 ODIR		=	obj
 SDIR		=	src
-LOGDIR		=	log_valgrind
+LOGDIR		=	logs
 # =============================== COMPILATION ================================ #
 CXX			=	c++
 CXXFLAGS	=	-std=c++98 -Wall -Wextra -Werror -MMD -MP
