+++ mk/options.mk
@@ -1,10 +1,7 @@
-MAKE	=	@make --no-print-directory
+MAKE		=	@make --no-print-directory
 VALGRIND	=	valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes --error-exitcode=1 --quiet --log-file=$(LOGDIR)/valgrind-%p.log
 
-# For parallel compilation
-# $(info $(MAKEFLAGS))
 JOBS := $(filter -j%,$(MAKEFLAGS))
-# $(info $(JOBS))
 
 ifeq ($(JOBS),)
     ifeq ($(shell uname -s), Darwin)
@@ -15,6 +12,3 @@ ifeq ($(JOBS),)
 endif
 
 FMAKE	=	$(MAKE) $(JOBS)
-
-# MAKEFLAGS += $(JOBS)
-# $(info $(JOBS))
